Improve handling of wiki links and backlinks
Substituting the links and finding backlinks is now done with a more
effienct approach. The algorithm‘s runtime is now linear in the number
of notes instead of quadratic.

Any occurrence of an ID like for example '20220218222506' (14 digits not
enclosed in double brackets) in a note is now treated as an ID of that
note. As before, the filename is also treated as an ID. All links using
such an ID will link to the correct note. Links via the title of a note
will not work anymore though.

Backlinks are now also displayed for literature notes.

Internal wiki links now look all the same: [◦]. This is because the text
of a wiki link is not really meaningful in Zettlr anyway. This way, no
space is wasted for 14-digit-long links, which should improve
readability.